Replace int by size_t for loop indices

+++ example/cxx/sparse/heisenberg_crs_mpi.cpp
@@ -50,7 +50,7 @@ int main(int argc, char *argv[]) {
     cols.clear();
     values.clear();
     double diag = 0;
-    for (int l = 0; l < lattice.size(); ++l) {
+    for (size_t l = 0; l < lattice.size(); ++l) {
       int i = lattice[l].first;
       int j = lattice[l].second;
       int m1 = 1 << i;
@@ -92,7 +92,7 @@ int main(int argc, char *argv[]) {
     for (int i = 0; i < num_conv; ++i) std::cout << ' ' << solver.eigenvalue(i);
     std::cout << std::endl;
     std::cout << "smallest eigenvector: ";
-    for (int j = 0; j < eigvec.size(); ++j) std::cout << eigvec[j] << ' ';
+    for (size_t j = 0; j < eigvec.size(); ++j) std::cout << eigvec[j] << ' ';
     std::cout << std::endl;
   }
